fix(non linear font scaling): qs customizer page tile height is not scaling with font size
Originally the tile view height in QS customizer page is fixed to 80dp, so when font size becomes larger the subtitle in tile might be cut off. Therefore, we adjust the tile view height to scale with the current font size to ensure the tile view height is enough for containing the content.

Bug: 287170018
Test: manually
Change-Id: I31c5479b64f71ad7bcf9a8688193e63a17d63251